refactor: index.tsx에서 이름 명시
