+++ src/pages/Community/CommunityDetail/CommunityDetail.tsx
