+++ src/pages/index.ts
@@ -1,2 +1,3 @@
-export { default as Main } from './Main';
-export { default as GardenEdit } from './Create/GardenEdit';
+export { default as Main } from './Main/Main';
+export { default as GardenEdit } from './Create/GardenEdit/GardenEdit';
+export { default as MyGardenEdit } from './Create/MyGardenEdit/MyGardenEdit';
