+++ src/routes/Layout.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@chakra-ui/react';
 import { Outlet } from 'react-router-dom';
-import Header from '@/components/Header';
+import { Header } from '@/components';
 
 const Layout = () => {
   return (
